requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Werkzeug==3.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==10.4.0
email-validator==2.2.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
gevent==24.2.1
greenlet==3.0.3
openpyxl==3.1.5
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
cachetools==5.5.0
tenacity==9.0.0
# Production scalability packages
celery==5.3.4
redis==5.0.1
Flask-Caching==2.1.0

